+++ NEWS.md
@@ -20,6 +20,7 @@ Unreleased
 * Introduce `suspenders:cleanup:organize_gemfile` task
 * Introduce `suspenders:production:environment` generator
 * Introduce `suspenders:test:environment` generator
+* Introduce `suspenders:development:environment` generator
 
 20230113.0 (January, 13, 2023)